// landing-page/src/types/SiteInfo.ts
export interface SiteInfoProps {
    id: number;

    // hero area
    banner_title: string;
    banner_text: string;

    //sessao 1
    section1_title: string;
    section1_text: string;

    //sessao 2
    section2_title: string;
    section2_text: string;

    //sessao 3
    section3_title: string;
    section3_text: string;

    //social network
    link_facebook: string;
    link_instagram: string;
    link_linkedin: string;

    // contact
    address: string;
    phone1: string;
    phone2: string;
    email: string;

    counters_1: number;
    counters_2: number;
    counters_3: number;
    counters_4: number;

    term_text: string;

    price_question: number;
}